🚀 change heartbeat timeout field type to float

+++ src/wechaty_puppet/schemas/event.py
@@ -132,4 +132,4 @@ class EventResetPayload(EventPayloadBase):
 @dataclass
 class EventHeartbeatPayload(EventPayloadBase):
     data: str
-    timeout: Optional[int] = None
+    timeout: Optional[float] = None
